third_party/WebKit/LayoutTests/vr/latest/navigator_vr_early_detached.html

<!DOCTYPE html>
<html>
<head>
<title>Detached use of navigator.vr()</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>
<body>
<iframe sandbox="allow-same-origin" id="subframe"></iframe>
<script>
//This test ensures that if a navigator.vr gets called on a detached navigator the renderer does not crash.
test(function() {
  var nav = window.frames[0].navigator;
  document.getElementById("subframe").remove();
  // If all goes well, nav.vr should just be null, otherwise the renderer will crash. 
  assert_equals(nav.vr, null);
}, "navigator.vr does not crash");
</script>
</body>
</html>
